Estacionamiento/requirements.txt
requests==2.31.0
mysql-connector-python==8.2.0
matplotlib==3.8.2
